requirements.txt
colorama
python-dotenv
python-telegram-bot
aiohttp
